fix everyone mention

+++ src/Commands/Fun/count.js
@@ -16,7 +16,8 @@ module.exports = {
 		countData.save();
 		const globalCount = await countSchema.findOne({ member: 'global' });
 		if (globalCount.number == 1500) return message.channel.send('**🤔 Maximum number in the counter already reached .**');
-		client.channels.cache.get('883724314426363994').send(!args.length ? `**• ${message.author} »** __${globalCount.number + 1}__ .` : `**• ${message.author} »** __${globalCount.number + 1}__ (${args.join(' ')}) .`).then(() => {
+		if (message.content.legnth > 20) return message.channel.send('**🤔 You can\'t send more than 20 character .**');
+		client.channels.cache.get('883724314426363994').send(!args.length ? `**• ${message.author} »** __${globalCount.number + 1}__ .` : `**• ${message.author} »** __${globalCount.number + 1}__ \`${args.join(' ')}\` .`).then(() => {
 			message.delete();
 			message.reply(`**You added one number to the counter, your score: (__${countData.number + 1}__)**`);
 			countData.number++;
